lesson 12 batchnorm

+++ dl2/scripts/clean_nb_output.py
@@ -18,9 +18,14 @@ def removeComments(inputFileName, outputFileName):
 if __name__ == "__main__":
     print(os.listdir())
     for f in os.listdir():
+        if os.path.isdir(f):
+            continue
         if f != 'clean_nb_output.py'and '_clean' not in f:
-            pre, post = f.split('.')
-            newfile = pre+'_clean.py'
-            removeComments(f,newfile)
-            os.remove(f)
-            os.rename(newfile,f)
+            try:
+                pre, post = f.split('.')
+                newfile = pre+'_clean.py'
+                removeComments(f,newfile)
+                os.remove(f)
+                os.rename(newfile,f)
+            except Exception as e:
+                print(f'Error with {f}: {e}')
